Some refactoring after comments
