+++ src/hotspot/share/services/memJfrReporter.hpp
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2022, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ *
+ */
+
+#ifndef SHARE_SERVICES_MEMJFRREPORTER_HPP
+#define SHARE_SERVICES_MEMJFRREPORTER_HPP
+
+#include "memory/allocation.hpp"
+#include "services/memBaseline.hpp"
+#include "utilities/globalDefinitions.hpp"
+#include "utilities/ticks.hpp"
+
+// The MemJFRReporter and MemJFRBaseline classes are only to be used from the
+// thread sending periodic JFR events. So no synchronization is needed.
+class MemJFRReporter : public AllStatic {
+private:
+  static void sendTypeEvent(const Ticks& starttime, const char* tag, size_t reserved, size_t committed);
+ public:
+  static void sendTotalEvent();
+  static void sendTypeEvents();
+};
+
+// Helper class to avoid taking multiple NMT baselines for
+// the two JFR events that are using the same data.
+class MemJFRBaseline : public AllStatic {
+private:
+  // The baseline age threshold in millie seconds. If older
+  // that this we will make a new baseline.
+  static const uint64_t BaselineAgeThreshold = 50;
+
+  static Ticks _baseline_timestamp;
+  static MemBaseline _baseline;
+
+public:
+  static MemBaseline& getBaseline();
+  static Ticks getTimestamp();
+};
+
+#endif //SHARE_SERVICES_MEMJFRREPORTER_HPP
